PycharmProjects/Travel Concierge Demo/backend/requirements.txt
# AI Travel Concierge Platform - Backend Dependencies (Windows Compatible)
# Essential packages that install without conflicts

# Core FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP and networking
requests>=2.31.0
httpx>=0.25.0
websockets>=12.0

# Environment and configuration
python-dotenv>=1.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Image processing (essential for image search)
Pillow>=10.0.0

# Data processing
pandas>=2.0.0
aiofiles>=23.0.0

# Basic AI/ML (optional, can be added later)
openai>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0 